Added better logging in server and authentication modules, introduced logging whitelists, improved configuration management, and enhanced error handling for authentication.

+++ src/shared/config.ts
@@ -10,11 +10,23 @@ type Config = {
   dbpassword: string | undefined;
   database: string | undefined;
   dbssl: string | undefined;
-  dbrejectunauthorized: boolean | undefined;
-  node_env: string | undefined;
+  dbrejectunauthorized: boolean;
+  log_level: string;
+  node_env: string;
 };
 
-const { APIPORT, DBPORT, DBHOST, DBUSER, DBPASSWORD, DATABASE, DBSSL, DBREJECTUNAUTHORIZED, NODE_ENV } = process.env;
+const {
+  APIPORT,
+  DBPORT,
+  DBHOST,
+  DBUSER,
+  DBPASSWORD,
+  DATABASE,
+  DBSSL,
+  DBREJECTUNAUTHORIZED,
+  LOG_LEVEL,
+  NODE_ENV
+} = process.env;
 
 const config: Config = {
   apiport: Number(APIPORT),
@@ -25,13 +37,14 @@ const config: Config = {
   database: DATABASE,
   dbssl: DBSSL,
   dbrejectunauthorized: DBREJECTUNAUTHORIZED === "true",
-  node_env: NODE_ENV
+  log_level: LOG_LEVEL ?? 'info',
+  node_env: NODE_ENV ?? 'production'
 };
 
 const verifyConfig = (conf: Config) => {
   for (const key in conf) {
     log.debug(`Checking ${key.toUpperCase()} in process.env`);
-    if (!(key.toUpperCase() in process.env) && key.toUpperCase() !== 'NODE_ENV') {
+    if (!(key.toUpperCase() in process.env) && key.toUpperCase() !== 'NODE_ENV' && key.toUpperCase() !== 'LOG_LEVEL') {
       log.error(`Missing Env Var ${key.toUpperCase()}`);
       process.kill(process.pid, 'SIGTERM');
     }
